Filled in the rest of the interface for the structural_sequence_labeling_trainer and cleaned up a few other related things.

+++ dlib/svm/cross_validate_sequence_labeler.h
@@ -42,7 +42,7 @@ namespace dlib
             for (unsigned long j = 0; j < pred.size(); ++j)
             {
                 const unsigned long truth = labels[i][j];
-                if (truth >= res.nr())
+                if (truth >= static_cast<unsigned long>(res.nr()))
                 {
                     // ignore labels the labeler doesn't know about.
                     continue;
